+++ resources/lang/en/challenges.php
@@ -44,6 +44,8 @@ return [
     'mark_text' => 'Are you sure you have done all the the work? Trust is needed there.',
 
     'comment' => 'Comment',
+    'show_challenge' => 'Show challenge',
+    'show_all' => 'Show all',
 
 
 ];