+++ resources/lang/es/challenges.php
@@ -44,5 +44,7 @@ return [
     'mark_text' => '¿Estás seguro/a que has superado la tarea? La confianza es necesaria.',
 
     'comment' => 'Comenta',
+    'show_challenge' => 'Muestra desafío',
+    'show_all' => 'Muestra todos',
 
 ];